% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfgs_mle.R
\name{bfgs_mle}
\alias{bfgs_mle}
\title{Calculate the MLE for beta using BFGS}
\usage{
bfgs_mle(design, outcome, model = "linear")
}
\arguments{
\item{design}{a n x p design matrix of numeric or factor predictors, possible including an intercept.}

\item{outcome}{a n-length numeric vector of outcomes.}

\item{model}{a character indicating the link function between the linear predictors and the mean of the outcome.}
}
\value{
A p-length vector, the estimated MLE for beta, i.e., value of beta that maximizes the (log) likelihood, as a vector.
}
\description{
Calculate the MLE for beta using BFGS
}
\details{
This function calculates the MLE for beta in a least squares regression
using BFGS (a quasi-Newton method using the gradient of the least squares function).
}
