% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.hiperglm}
\alias{coef.hiperglm}
\title{Estimated coefficients of hiperglm object}
\usage{
\method{coef}{hiperglm}(hiperglm)
}
\arguments{
\item{hiperglm}{a fitted hiperglm object.}
}
\description{
Estimated coefficients of hiperglm object
}
\details{
This function returns the estimated regression coefficients of a fitted hiperglm object.
}
