% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_functions.R
\name{gradient_of_log_lik}
\alias{gradient_of_log_lik}
\title{Gradient of log likelihood of regression model at a point}
\usage{
gradient_of_log_lik(
  design,
  outcome,
  beta,
  model = "linear",
  Sigma_inv = diag(length(outcome))
)
}
\arguments{
\item{design}{a n x p design matrix of numeric or factor predictors, possibly including an intercept.}

\item{outcome}{a n-length numeric vector of outcomes.}

\item{beta}{a p-length numeric vector at which to evaluate the log likelihood, possibly including an intercept.}

\item{model}{a character indicating the link function between the linear predictors and the mean of the outcome.}

\item{Sigma_inv}{the covariance matrix for the errors; default is nxn identity matrix; doesn't influence MLE estimate of beta (though would influence estimate of variance of beta).}
}
\value{
a p x 1 matrix, the gradient of the log likelihood of the data at beta.
}
\description{
Gradient of log likelihood of regression model at a point
}
\details{
This function computes, analytically, the gradient of the log likelihood in linear regression.
}
