% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_functions.R
\name{log_lik}
\alias{log_lik}
\title{Log likelihood of regression model at a point}
\usage{
log_lik(
  design,
  outcome,
  beta,
  model = "linear",
  Sigma_inv = diag(length(outcome))
)
}
\arguments{
\item{design}{a n x p design matrix of numeric or factor predictors, possibly including an intercept.}

\item{outcome}{a n-length numeric vector of outcomes.}

\item{beta}{a p-length numeric vector at which to evaluate the log likelihood, possibly including an intercept.}

\item{model}{a character indicating the link function between the linear predictors and the mean of the outcome.}

\item{Sigma_inv}{the covariance matrix for the errors; default is nxn identity matrix; doesn't influence MLE estimate of beta (though would influence estimate of variance of beta).}
}
\value{
A scalar, the log likelihood of the data at beta.
}
\description{
Log likelihood of regression model at a point
}
\details{
This function computes the log likelihood of data in a linear regression model at a given value of beta.
}
