% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_mle.R
\name{newton_mle}
\alias{newton_mle}
\title{Calculate the MLE for beta using Newton's method}
\usage{
newton_mle(
  design,
  outcome,
  model = "logistic",
  abs_conv_thres = 1e-06,
  rel_conv_thres = 1e-06,
  max_iter = 200
)
}
\arguments{
\item{design}{a n x p design matrix of numeric or factor predictors, possible including an intercept.}

\item{outcome}{a n-length numeric vector of outcomes.}

\item{model}{a character indicating the link function between the linear predictors and the mean of the outcome.}

\item{abs_conv_thres}{a double indicating the absolute convergence threshold for change in log likelihood; note that likelihood ratio under null beta tells us we can get a 0.5 increase in log likelihood by chance}

\item{rel_conv_thres}{a double indicating the relative convergence threshold for change in log likelihood}

\item{max_iter}{an integer indicating the maximum number of iterations to run before giving up on convergence (if convergence threshold hasn't been met)}
}
\value{
The estimated MLE for beta, i.e., value of beta that maximizes the (log) likelihood, as a vector.
}
\description{
Calculate the MLE for beta using Newton's method
}
\details{
This function calculates the MLE for beta in a least squares regression
using Newton's method (using the gradient and Hessian of the least squares function).
}
