% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.hiperglm}
\alias{print.hiperglm}
\title{Printed summary of hiperglm object}
\usage{
\method{print}{hiperglm}(x, ...)
}
\arguments{
\item{x}{a fitted hiperglm object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Printed summary of hiperglm object
}
\details{
This function returns prints summary characteristics of a fitted hiperglm object.
}
