% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoinverse_mle.R
\name{pseudoinverse_mle}
\alias{pseudoinverse_mle}
\title{Calculate the MLE for beta in OLS using pseudoinverse of design matrix}
\usage{
pseudoinverse_mle(design, outcome)
}
\arguments{
\item{design}{a n x p design matrix of numeric or factor predictors, possible including an intercept.}

\item{outcome}{a n-length numeric vector of outcomes.}
}
\value{
The estimated MLE for beta, i.e., (X^T X)^{-1} X^T y, as a vector.
}
\description{
Calculate the MLE for beta in OLS using pseudoinverse of design matrix
}
\details{
This function calculates the MLE for beta in a linear regression
using the pseudoinverse of the design matrix.
}
