% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.hiperglm}
\alias{vcov.hiperglm}
\title{Estimated variance-covariance matrix of hiperglm object}
\usage{
\method{vcov}{hiperglm}(hiperglm)
}
\arguments{
\item{hiperglm}{a fitted hiperglm object.}
}
\description{
Estimated variance-covariance matrix of hiperglm object
}
\details{
This function returns the estimated variance-covariance matrix of a fitted hiperglm object.
}
